+++ src/api/v2/order.rs
@@ -517,7 +517,7 @@ impl ChangeReqInit {
 
 
 /// A PATCH request to be made to the /v2/orders/<order-id> endpoint.
-#[derive(Clone, Debug, Serialize, PartialEq)]
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct ChangeReq {
   /// Number of shares to trade.
   #[serde(rename = "qty")]
@@ -982,7 +982,7 @@ mod tests {
 
   /// Check that we can serialize and deserialize an [`OrderReq`].
   #[test]
-  fn serialize_deserialize_request() {
+  fn serialize_deserialize_order_request() {
     let request = OrderReqInit {
       type_: Type::TrailingStop,
       trail_price: Some(Num::from(50)),
@@ -992,6 +992,21 @@ mod tests {
 
     let json = to_json(&request).unwrap();
     assert_eq!(from_json::<OrderReq>(&json).unwrap(), request);
+  }
+
+  /// Check that we can serialize and deserialize a [`ChangeReq`].
+  #[test]
+  fn serialize_deserialize_change_request() {
+    let request = ChangeReqInit {
+      quantity: Num::from(37),
+      time_in_force: TimeInForce::UntilCanceled,
+      trail: Some(Num::from(42)),
+      ..Default::default()
+    }
+    .init();
+
+    let json = to_json(&request).unwrap();
+    assert_eq!(from_json::<ChangeReq>(&json).unwrap(), request);
   }
 
   /// Verify that we can submit a limit order.
@@ -1267,6 +1282,8 @@ mod tests {
     };
   }
 
+  /// Check that we get back the expected error when attempting to
+  /// cancel an invalid (non-existent) order.
   #[test(tokio::test)]
   async fn cancel_invalid_order() {
     let id = Id(Uuid::parse_str("00000000-0000-0000-0000-000000000000").unwrap());
@@ -1281,6 +1298,7 @@ mod tests {
     };
   }
 
+  /// Check that we can retrieve an order given its ID.
   #[test(tokio::test)]
   async fn retrieve_order_by_id() {
     let api_info = ApiInfo::from_env().unwrap();
@@ -1338,6 +1356,7 @@ mod tests {
     };
   }
 
+  /// Check that we can change an existing order.
   #[test(tokio::test)]
   async fn change_order() {
     let request = OrderReqInit {