+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 - Added support for subscribing to realtime trades
 - Added `class` member to `api::v2::order::Order` type
 - Added `symbols` member to `api::v2::orders::OrdersReq` type
+- Added `Deserialize` implementation for `api::v2::order::ChangeReq`
 - Bumped `uuid` dependency to `1.0`